Fix activesupport require error
The right hand column showing the /feed route not working. On Heroku JS console
is showing...

"Failed to load resource: the server responded with a status of 500 (Internal Server Error)"

In the log showing...

"NameError - uninitialized constant ActiveSupport::XMLConverter::XmlMini"

Seems to be related to a change made to activesupport that now needs
'require "activesupport"'.