Display data for people currently on GOV.UK
Display number of users currently on GOV.UK. The number has been hardcoded but will be switched to use data from the GA realtime API which is being completed as a separate task.

Display data for the number of people on GOV.UK over the last 24  hours in a graph. Currently the data is hard coded, and will use data from the GA API which is being completed as a separate task.

Co-authored-by: Graham Lewis <graham.lewis@digital.cabinet-office.gov.uk>