Use GA4 data as opposed to UA data for /popular-content
Updates API client from Universal Analytics API (AnalyticsreportingV4::AnalyticsReportingService) to
GA4 API (Google::Analytics::Data::V1beta::AnalyticsData::Client).

We are in the process of switching from UA to GA4, but page views has been implemented for most of
GOV.UK. This and an issue with our Realtime UA data being unreliable, means that we are
now using the GA4 API.

This also means that the same credentials can be used for /popular-content, /realtime-traffic
and /traffic-report.