Handle API errors
Ensure the display screen doesn't look too horrible of there is an API error!
Will display 'No data available' if there is an error.

I did experiment with using this [error reporting gem](https://cloud.google.com/error-reporting/docs/setup/ruby).
It did report the errors but it stops the server even with a rescue block. I thought the standard error
reporting was still informative and this approach also allows us to still display something nice to the user,
rather than a crash screen.

You can test this by changing to property in the request to a random number.