Added chart with api data
The chart is now using api data and labels.
We have excluded the total number of users in the past 24 hours
from the loop for the chart as it didn't relate to the main purpose of the chart. We may
wish to display it somewhere else however, perhaps near the realtime ticker.