Fix incorrectly named value
Previously I formatted the '(other)' row that came back from the API as `total_number_of_users`. I incorrectly thought this was what the value represented. I've since learnt that:

When the result of multiplying the cardinality of each dimension in a report exceeds the row limit, Analytics displays an "(other)" row.

https://support.google.com/analytics/answer/9309767?hl=en#zippy=%2Cin-this-article

Therefore I am removing the method that incorrectly formatted the name of the 'other' hour value.

We are currently working on eliminating or reducing this 'other' value.