Set row offset and limit
This will just return 25 rows. We were advised to set this in order to try and reduce cardinality issues. The '(other)' value is no longer returned as a result of setting this, which seems positive. However, the numbers for the other hour values remain the same as when it isn't set...