Change graph to display 24 hours of data relative to current hour
Previously we were using the `hour` dimension with the date range set to `yesterday` -> `today`. This was actually giving us data for 2 days summed into 1. See PR description for screen grabs of the generated graph.

Using the `hour` value would be better if we wanted data from only today (so far) or the whole of only yesterday. Using dateHour allows us to display data for the last 24 hours relative to the current hour.

The AJAX request has been set to update once an hour. This should mean that the graph updates over the course of the day, which will be more interesting to look at.

Some formatting has been done on the labels to just show the hour values.

There is still a data issue with the numbers themselves due to cardinality. We are going to speak to Google about this.